requirements.txt
accelerate==0.21.0
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
bitsandbytes==0.41.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
dataclasses-json==0.5.14
debugpy==1.8.0
decorator==5.1.1
dnspython==2.4.2
docopt==0.6.2
einops==0.6.1
exceptiongroup==1.1.3
executing==2.0.0
filelock==3.12.4
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
greenlet==3.0.0
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.4.0
jupyter_core==5.4.0
kiwisolver==1.4.5
langchain==0.0.240
langsmith==0.0.44
loguru==0.7.2
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.1
openapi-schema-pydantic==1.2.4
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pinecone-client==2.2.2
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pydantic==1.10.13
pydantic_core==2.10.1
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.31.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
wcwidth==0.2.8
xformers==0.0.20
yarg==0.1.9
yarl==1.9.2
zipp==3.17.0
